+++ src/design/plone/contenttypes/restapi/serializers/documento.py
@@ -1,26 +1,23 @@
 # -*- coding: utf-8 -*-
-from .related_news_serializer import (
-    SerializeFolderToJson as RelatedNewsSerializer,
+from Acquisition import aq_inner
+from design.plone.contenttypes.interfaces.documento import IDocumento
+from design.plone.contenttypes.restapi.serializers.dxcontent import (
+    SerializeFolderToJson,
 )
-from design.plone.contenttypes.interfaces.documento import (
-    IDocumento,
-)
-
 from plone.restapi.interfaces import ISerializeToJson, ISerializeToJsonSummary
+from zc.relation.interfaces import ICatalog
 from zope.component import adapter, getMultiAdapter
+from zope.component import getUtility
+from zope.globalrequest import getRequest
 from zope.interface import implementer
 from zope.interface import Interface
-from Acquisition import aq_inner
-from zope.component import getUtility
 from zope.intid.interfaces import IIntIds
 from zope.security import checkPermission
-from zc.relation.interfaces import ICatalog
-from zope.globalrequest import getRequest
 
 
 @implementer(ISerializeToJson)
 @adapter(IDocumento, Interface)
-class DocumentoSerializer(RelatedNewsSerializer):
+class DocumentoSerializer(SerializeFolderToJson):
     def get_services(self):
         """
         """
@@ -49,7 +46,8 @@ class DocumentoSerializer(RelatedNewsSerializer):
         return sorted(services, key=lambda k: k["title"])
 
     def __call__(self, version=None, include_items=True):
-        self.index = "news_uo"
+        if "b_size" not in self.request.form:
+            self.request.form["b_size"] = 200
         result = super(DocumentoSerializer, self).__call__(
             version=version, include_items=include_items
         )