+++ CHANGES.rst
@@ -7,6 +7,8 @@ Changelog
 
 - Added backreferences to Documento and Cartella Modulistica for related services.
   [daniele]
+- Documento now set b_size=200 by default to show more than 25 items when getting its data.
+  [cekk]
 
 
 3.0.1 (2021-05-04)